Cast study time sums to bigint in admin stats functions
sum(bigint) returns numeric in Postgres, so both admin RPC functions
failed at runtime with 42804 (structure of query does not match
function result type). Cast the coalesced sums back to bigint to match
the declared return columns. Already applied to production.

+++ supabase/migrations/20260612_admin_dashboard_stats.sql
@@ -27,7 +27,8 @@ begin
     (select count(distinct ss.user_id)
        from public.study_sessions ss
       where ss.created_at >= p_day_start),
-    (select coalesce(sum(ss.duration), 0) from public.study_sessions ss),
+    -- sum(bigint) returns numeric; cast back to match the declared bigint column
+    (select coalesce(sum(ss.duration), 0)::bigint from public.study_sessions ss),
     (select count(*) from public.profiles p where p.created_at >= p_day_start);
 end;
 $$;
@@ -51,7 +52,7 @@ begin
   end if;
 
   return query
-  select coalesce(sum(ss.duration), 0), count(*)
+  select coalesce(sum(ss.duration), 0)::bigint, count(*)
   from public.study_sessions ss
   where ss.user_id = p_user_id;
 end;
